refactor components and page
